requirements.txt
behave>=1.2.6
selenium>=4.15.2
webdriver-manager>=4.0.1
requests>=2.31.0
pytest>=7.4.3
allure-behave>=2.13.2
python-dotenv>=1.0.0
Pillow>=10.4.0
faker>=20.1.0
